+++ Tests/Unit/Domain/Model/Event/SingleEventTest.php
@@ -1896,4 +1896,23 @@ final class SingleEventTest extends UnitTestCase
     {
         self::assertFalse($this->subject->isEventTopic());
     }
+
+    /**
+     * @test
+     */
+    public function getBillingStartByDefaultReturnsNull(): void
+    {
+        self::assertNull($this->subject->getBillingStart());
+    }
+
+    /**
+     * @test
+     */
+    public function setBillingStartSetsBillingStart(): void
+    {
+        $date = new \DateTime();
+        $this->subject->setBillingStart($date);
+
+        self::assertSame($date, $this->subject->getBillingStart());
+    }
 }